Add support langchain stream
When calling the stream method in LangChain, the on_chain_start method takes a dummy input: {"input":""}. This is because if the input is a streaming data, it needs to collect all the streaming data completely before executing the update operation in the on_chain_end or on_chain_error callback methods.

In the on_chain_end and on_chain_error methods, when determining if the input kwargs has the inputs parameter, if it does, the inputs parameter is passed during the update trace execution to achieve the effect of updating the span's input.