fix: removing the run_id from runs
Once the runs are generated in the current scenario when the callback is declared as a single declaration and re-used, then the runs basically starts collecting the run traces indefinitely untill it goes OOM. To avoid this situtaion we remove the run details after the end of the chain and at the end of the run invocation itself. This ensures the details are persistent till the end of the loop and is removed once the traces are pushed to langfuse.