Do another update call when we're updating an ended trace in update_current_trace
It's possible to have observations that 'outlive' their trace, for
example when returning a StreamingResponse in FastAPI.

Currently, calling update_current_trace in such cases only worked
for the server_merged_attributes. We can work around this
by sending the metadata in .trace() instead of storing it in the context.